requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
pydantic-settings==2.13.1
anthropic==0.84.0
faiss-cpu==1.13.2
sentence-transformers==5.2.3
python-pptx==1.0.2
pandas==2.2.3
openpyxl==3.1.5
aiofiles==23.2.1
httpx==0.27.2
python-dotenv==1.0.1
